requirements.txt
easyocr
opencv-python-headless
Pillow
numpy
torch
torchvision
